Add debugging properties and update condition in props
The changes in `Directory.Build.props` add a new `<PropertyGroup>`
that includes properties for debugging and source embedding.
Specifically, the properties `DebugSymbols`, `DebugType`, and
`EmbedAllSources` are set to `true`, `portable`, and `true`
respectively.

The changes in `Project.Common.props` modify the condition for a
`<PropertyGroup>`. The original condition checked if the project
was being built inside Visual Studio or by ReSharper. The new
condition checks if the `UseProjectReference` property is set to
`true` instead.